Change .NET to 4.5, some changes

+++ Coob/CSystem/Coob.cs
@@ -27,7 +27,6 @@ namespace Coob
             MaxClients = 1024;
         }
     }
-
     public class Coob
     {
         public ConcurrentQueue<Packet.Base> MessageQueue;
@@ -183,7 +182,7 @@ namespace Coob
 
             foreach (var client in GetClients())
             {
-                client.Writer.Write(SCPacketIDs.ServerChatMessage);
+                client.Writer.Write((int)SCPacketIDs.ServerChatMessage);
                 client.Writer.Write(id);
                 client.Writer.Write(msgLength);
                 client.Writer.Write(msgBuffer);
